+++ eng/tox/prep_sphinx_env.py
@@ -20,6 +20,7 @@ import ast
 import os
 import textwrap
 import io
+from tox_helper_tasks import get_package_details
 
 logging.getLogger().setLevel(logging.INFO)
 
@@ -115,41 +116,6 @@ def create_index(doc_folder, source_location, package_name):
     # write index
     with open(content_destination, "w+") as f:
         f.write(index_content)
-
-
-def get_package_details(setup_filename):
-    mock_setup = textwrap.dedent(
-        """\
-    def setup(*args, **kwargs):
-        __setup_calls__.append((args, kwargs))
-    """
-    )
-    parsed_mock_setup = ast.parse(mock_setup, filename=setup_filename)
-    with io.open(setup_filename, "r", encoding="utf-8-sig") as setup_file:
-        parsed = ast.parse(setup_file.read())
-        for index, node in enumerate(parsed.body[:]):
-            if (
-                not isinstance(node, ast.Expr)
-                or not isinstance(node.value, ast.Call)
-                or not hasattr(node.value.func, "id")
-                or node.value.func.id != "setup"
-            ):
-                continue
-            parsed.body[index:index] = parsed_mock_setup.body
-            break
-
-    fixed = ast.fix_missing_locations(parsed)
-    codeobj = compile(fixed, setup_filename, "exec")
-    local_vars = {}
-    global_vars = {"__setup_calls__": []}
-    current_dir = os.getcwd()
-    working_dir = os.path.dirname(setup_filename)
-    os.chdir(working_dir)
-    exec(codeobj, global_vars, local_vars)
-    os.chdir(current_dir)
-    _, kwargs = global_vars["__setup_calls__"][0]
-
-    return kwargs["name"], kwargs["version"]
 
 
 def write_version(site_folder, version):